[AIRFLOW-5743] Move Google PubSub to providers package
