+++ airflow/contrib/hooks/gcp_pubsub_hook.py
@@ -16,14 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.gcp.hooks.pubsub`."""
+"""This module is deprecated. Please use `airflow.providers.google.cloud.hooks.pubsub`."""
 
 import warnings
 
 # pylint: disable=unused-import
-from airflow.gcp.hooks.pubsub import PubSubException, PubSubHook  # noqa
+from airflow.providers.google.cloud.hooks.pubsub import PubSubException, PubSubHook  # noqa
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.gcp.hooks.pubsub`.",
+    "This module is deprecated. Please use `airflow.providers.google.cloud.hooks.pubsub`.",
     DeprecationWarning, stacklevel=2
 )
