+++ airflow/contrib/operators/pubsub_operator.py
@@ -16,17 +16,17 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.gcp.operators.pubsub`."""
+"""This module is deprecated. Please use `airflow.providers.google.cloud.operators.pubsub`."""
 
 import warnings
 
 # pylint: disable=unused-import
-from airflow.gcp.operators.pubsub import (  # noqa
+from airflow.providers.google.cloud.operators.pubsub import (  # noqa
     PubSubPublishOperator, PubSubSubscriptionCreateOperator, PubSubSubscriptionDeleteOperator,
     PubSubTopicCreateOperator, PubSubTopicDeleteOperator,
 )
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.gcp.operators.pubsub`.",
+    "This module is deprecated. Please use `airflow.providers.google.cloud.operators.pubsub`.",
     DeprecationWarning, stacklevel=2
 )
