+++ airflow/providers/google/cloud/operators/pubsub.py
@@ -25,8 +25,8 @@ from typing import Dict, List, Optional, Sequence, Tuple, Union
 from google.api_core.retry import Retry
 from google.cloud.pubsub_v1.types import Duration, MessageStoragePolicy, PushConfig
 
-from airflow.gcp.hooks.pubsub import PubSubHook
 from airflow.models import BaseOperator
+from airflow.providers.google.cloud.hooks.pubsub import PubSubHook
 from airflow.utils.decorators import apply_defaults
