+++ tests/providers/google/cloud/sensors/test_pubsub.py
@@ -23,7 +23,7 @@ from google.cloud.pubsub_v1.types import ReceivedMessage
 from google.protobuf.json_format import MessageToDict, ParseDict
 
 from airflow.exceptions import AirflowSensorTimeout
-from airflow.gcp.sensors.pubsub import PubSubPullSensor
+from airflow.providers.google.cloud.sensors.pubsub import PubSubPullSensor
 from tests.compat import mock
 
 TASK_ID = 'test-task-id'
@@ -50,14 +50,14 @@ class TestPubSubPullSensor(unittest.TestCase):
     def _generate_dicts(self, count):
         return [MessageToDict(m) for m in self._generate_messages(count)]
 
-    @mock.patch('airflow.gcp.sensors.pubsub.PubSubHook')
+    @mock.patch('airflow.providers.google.cloud.sensors.pubsub.PubSubHook')
     def test_poke_no_messages(self, mock_hook):
         operator = PubSubPullSensor(task_id=TASK_ID, project_id=TEST_PROJECT,
                                     subscription=TEST_SUBSCRIPTION)
         mock_hook.return_value.pull.return_value = []
         self.assertEqual([], operator.poke(None))
 
-    @mock.patch('airflow.gcp.sensors.pubsub.PubSubHook')
+    @mock.patch('airflow.providers.google.cloud.sensors.pubsub.PubSubHook')
     def test_poke_with_ack_messages(self, mock_hook):
         operator = PubSubPullSensor(task_id=TASK_ID, project_id=TEST_PROJECT,
                                     subscription=TEST_SUBSCRIPTION,
@@ -72,7 +72,7 @@ class TestPubSubPullSensor(unittest.TestCase):
             ack_ids=['1', '2', '3', '4', '5']
         )
 
-    @mock.patch('airflow.gcp.sensors.pubsub.PubSubHook')
+    @mock.patch('airflow.providers.google.cloud.sensors.pubsub.PubSubHook')
     def test_execute(self, mock_hook):
         operator = PubSubPullSensor(
             task_id=TASK_ID,
@@ -92,7 +92,7 @@ class TestPubSubPullSensor(unittest.TestCase):
         )
         self.assertEqual(generated_dicts, response)
 
-    @mock.patch('airflow.gcp.sensors.pubsub.PubSubHook')
+    @mock.patch('airflow.providers.google.cloud.sensors.pubsub.PubSubHook')
     def test_execute_timeout(self, mock_hook):
         operator = PubSubPullSensor(task_id=TASK_ID, project_id=TEST_PROJECT,
                                     subscription=TEST_SUBSCRIPTION,
